feat(NotificationPermission): 알림 관련 로직을 NotificationPermission.kt으로 분리, 알림 권한 요청을 메인 액티비티가 아닌 완료 액티비티에서 진행
