+++ app/src/main/java/woowacourse/movie/ui/main/MainActivity.kt
@@ -1,38 +1,18 @@
 package woowacourse.movie.ui.main
 
-import android.Manifest
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.content.pm.PackageManager
-import android.os.Build
 import android.os.Bundle
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import woowacourse.movie.R
-import woowacourse.movie.ui.seat.NotiChannel
-import woowacourse.movie.util.shortToast
 
 class MainActivity : AppCompatActivity() {
-    private val requestPermissionLauncher =
-        registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted ->
-            if (!isGranted) {
-                shortToast(R.string.permission_denied)
-            } else {
-                shortToast(R.string.permission_granted)
-            }
-        }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         initNavigation()
-        requestNotificationPermission()
-        createChannel()
     }
 
     private fun initNavigation() {
@@ -57,32 +37,5 @@ class MainActivity : AppCompatActivity() {
             replace(R.id.fragmentContainerView, fragment)
         }
         return true
-    }
-
-    private fun requestNotificationPermission() {
-        if (ContextCompat.checkSelfPermission(
-                this,
-                Manifest.permission.POST_NOTIFICATIONS,
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                if (!shouldShowRequestPermissionRationale(Manifest.permission.POST_NOTIFICATIONS)) {
-                    requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
-                }
-            }
-        }
-    }
-
-    private fun createChannel() {
-        val mChannel = NotificationChannel(
-            NotiChannel.BOOKING_ALARM.channelName,
-            NotiChannel.BOOKING_ALARM.channelName,
-            NotificationManager.IMPORTANCE_HIGH,
-        ).apply {
-            description = NotiChannel.BOOKING_ALARM.channelDescription
-        }
-        val notificationManager =
-            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        notificationManager.createNotificationChannel(mChannel)
     }
 }
